Move Plotly functionality to non-base class
